Further extract components out of the InfobyteEditor
