+++ site/src/Infobytes/organisms/InfobitsInput.svelte
@@ -1,0 +1,56 @@
+<script lang="ts">
+    import InfoBitEditor from "../templates/InfoBitEditor.svelte";
+    import { Card, Button } from "sveltestrap";
+    import { Infobit } from "../../stores";
+
+    export let infobits: any[];
+    export let errorInfobits: any;
+
+    const addInfobit = () => {
+        infobits = infobits.concat(new Infobit());
+        errorInfobits = errorInfobits.concat(new Infobit());
+    };
+
+    const removeInfobit = (i) => () => {
+        infobits = infobits.filter((_, j) => j !== i);
+        errorInfobits = errorInfobits.filter((_, j) => j !== i);
+    };
+</script>
+
+{#if infobits.length !== 0}
+    {#each infobits as infobit, i}
+        <InfoBitEditor bind:value={infobit} />
+        <div class="form-control-row">
+            {#if i === infobits.length - 1}
+                <button
+                    type="button"
+                    class="form-control-button"
+                    on:click={addInfobit}>+</button
+                >
+            {/if}
+            {#if infobits.length !== 1}
+                <button
+                    type="button"
+                    class="form-control-button"
+                    on:click={removeInfobit(i)}>-</button
+                >
+            {/if}
+        </div>
+    {/each}
+{:else}
+    <Card body
+        >Klicke auf 'Infobit hinzufügen +' um ein Infobit hinzu zufügen!</Card
+    >
+{/if}
+{#if infobits.length === 0}
+    <Button type="button" class="form-control-button" on:click={addInfobit}
+        >Infobit hinzufügen +</Button
+    >
+{/if}
+
+<style>
+    .form-control-button {
+        min-width: auto;
+        margin-right: 1em;
+    }
+</style>
